requirements.txt
selenium>=4.15.0
undetected-chromedriver>=3.5.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.12
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
